Create README.md
This experiment examines Fourier series expansion and finite series reconstruction of a
signal using MATLAB. This is accomplished through an illustrative MATLAB GUI and various
MATLAB functions. A list of these functions and their descriptions in the context they are used
here are given in Table 1 as a reference. It is suggested that at least the main functions are well
understood in order to carry out the experiment. Aside from that, the necessary files for the
MATLAB GUI will be provided to you and how it works will be explained below